requirements.txt
streamlit>=1.22.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
networkx>=3.0
matplotlib>=3.7.0
pandas>=2.0.0
pdfkit>=1.0.0
jinja2>=3.1.2
regex>=2022.10.31
numpy>=1.24.0
pyvis>=0.3.2
